requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
configobj==5.0.6
configparser==5.3.0
Django==3.2.16
djangorestframework==3.14.0
etelemetry==0.3.0
exceptiongroup==1.0.1
filelock==3.8.0
fitz==0.0.1.dev2
future==0.18.2
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
isodate==0.6.1
jsonify==0.5
looseversion==1.0.2
lxml==4.9.1
networkx==2.6.3
nibabel==4.0.2
nipype==1.8.5
numpy==1.21.6
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pathlib==1.0.1
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.21.0
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
pyxnat==1.5
rdflib==6.2.0
requests==2.28.1
scipy==1.7.3
selenium==4.6.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
traits==6.3.2
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.12
wsproto==1.2.0
zipp==3.10.0
opencv-python
matplotlib
pdf2image
Pillow
